requirements.txt
asgiref==3.6.0
cfgv==3.3.1
distlib==0.3.6
Django==4.2
django-phonenumber-field==7.1.0
filelock==3.12.0
identify==2.5.22
nodeenv==1.7.0
phonenumbers==8.13.11
platformdirs==3.2.0
pre-commit==3.2.2
PyYAML==6.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.22.0
